added a separate file for the polygon json files
