--- src/ch5/2d-incremental-collision-checking/octagon.json
@@ -1,13 +1,0 @@
-{
-  "points": [
-    {"x":-100,"y":200},
-    {"x":100,"y":200},
-    {"x":200,"y":100},
-    {"x":200,"y":-100},
-    {"x":100,"y":-200},
-    {"x":-100,"y":-200},
-    {"x":-200,"y":-100},
-    {"x":-200,"y":100}
-  ],
-  "origin": [{"x":0,"y":0}]
-  }
