+++ src/ch5/2d-incremental-collision-checking/polygon_loader.py
