+++ src/ch5/2d-incremental-collision-checking/polygons/boat.json
@@ -1,0 +1,8 @@
+{
+  "points": [
+    {"x":440, "y":530},
+    {"x":470,"y":590},
+    {"x":350,"y":530},
+    {"x":380,"y":470}
+  ]
+}
