+++ src/ch5/2d-incremental-collision-checking/polygons/displaced_rectangle.json
@@ -1,0 +1,8 @@
+{
+"points":[
+    {"x":650,"y":650},
+    {"x":830,"y":650},
+    {"x":830,"y":770},
+    {"x":650,"y":770}
+  ]
+}
