+++ src/ch5/2d-incremental-collision-checking/polygons/octagon.json
@@ -1,0 +1,13 @@
+{
+  "points": [
+    {"x":150,"y":450},
+    {"x":350,"y":450},
+    {"x":450,"y":350},
+    {"x":450,"y":150},
+    {"x":350,"y":50},
+    {"x":150,"y":50},
+    {"x":50,"y":150},
+    {"x":50,"y":350}
+  ],
+  "origin": [{"x":0,"y":0}]
+  }
