+++ src/ch5/2d-incremental-collision-checking/polygons/offset_triangle.json
