+++ src/ch5/2d-incremental-collision-checking/polygons/rectangle.json
@@ -6,3 +6,4 @@
     {"x":590,"y":710}
   ]
 }
+
